Adjust sources/tree to build with Bazel
