+++ .bazelversion
@@ -1,0 +1,1 @@
+5.1.1
