+++ .gitignore
@@ -1,2 +1,3 @@
-build/
+/build
 *.pyc
+/bazel-*
