+++ base/filesystem.h
@@ -23,6 +23,7 @@
 #ifndef __FILESYSTEM_UTIL_H__
 #define __FILESYSTEM_UTIL_H__
 
+#include <cstdint>
 #include <string>
 #include <vector>
 
@@ -129,7 +130,7 @@ enum fileTypes
 /**
  * Return a sorted list of the files in the specified directory.  listDir() will glob files from
  * the specified path, and filter against wildcard characters including `*` and `?`.
- * For example, valid paths would include `~/workspace`, `~/workspace/*.jpg`, ect.
+ * For example, valid paths would include `~/workspace`, `~/workspace/ *.jpg`, ect.
  *
  * @see here for a description of wildcard matching:  https://www.man7.org/linux/man-pages/man7/glob.7.html
  *
