+++ base/rand.h
