+++ camera/v4l2-console/BUILD
@@ -1,0 +1,16 @@
+cc_binary(
+    name = "v4l2_console",
+    srcs = [
+        "v4l2-console.cpp",
+    ],
+    linkopts = [
+        "-lGL",
+        "-lGLEW",
+        "-lGLU",
+        "-lX11",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//camera",
+    ],
+)
