+++ input/BUILD
@@ -1,0 +1,17 @@
+cc_library(
+    name = "input",
+    srcs = [
+        "devInput.cpp",
+        "devJoystick.cpp",
+        "devKeyboard.cpp",
+    ],
+    hdrs = [
+        "devInput.h",
+        "devJoystick.h",
+        "devKeyboard.h",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//base",
+    ],
+)
