+++ video/video-viewer/BUILD
@@ -1,0 +1,11 @@
+cc_binary(
+    name = "video_viewer",
+    srcs = ["video-viewer.cpp"],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//camera",
+        "//network",
+        "//video:video_output_create",
+        "//video:video_source_create",
+    ],
+)
