+++ video/videoSource.cpp
@@ -24,10 +24,14 @@
 #include "imageLoader.h"
 
 #include "gstCamera.h"
-#include "gstDecoder.h"
+// #include "gstDecoder.h"
 
 #include "logging.h"
 
+#include <cassert>
+#include <cstdint>
+#include <string>
+#include <algorithm>
 
 // constructor
 videoSource::videoSource( const videoOptions& options ) : mOptions(options)
@@ -44,79 +48,6 @@ videoSource::~videoSource()
 
 }
 
-// Create
-videoSource* videoSource::Create( const videoOptions& options )
-{
-	videoSource* src = NULL;
-	
-	const URI& uri = options.resource;
-
-	if( uri.protocol == "file" )
-	{
-		if( gstDecoder::IsSupportedExtension(uri.extension.c_str()) )
-			src = gstDecoder::Create(options);
-		else
-			src = imageLoader::Create(options);
-	}
-	else if( uri.protocol == "rtp" || uri.protocol == "rtsp" || uri.protocol == "webrtc" )
-	{
-		src = gstDecoder::Create(options);
-	}
-	else if( uri.protocol == "csi" || uri.protocol == "v4l2" )
-	{
-		src = gstCamera::Create(options);
-	}
-	else
-	{
-		LogError(LOG_VIDEO "videoSource -- unsupported protocol (%s)\n", uri.protocol.size() > 0 ? uri.protocol.c_str() : "null");
-	}
-
-	if( !src )
-		return NULL;
-
-	LogSuccess(LOG_VIDEO "created %s from %s\n", src->TypeToStr(), src->GetResource().string.c_str());
-	src->GetOptions().Print(src->TypeToStr());
-	return src;
-}
-
-// Create
-videoSource* videoSource::Create( const char* resource, const commandLine& cmdLine, int positionArg, const videoOptions& options )
-{
-	videoOptions opt = options;
-
-	if( !opt.Parse(resource, cmdLine, videoOptions::INPUT, positionArg) )
-	{
-		LogError(LOG_VIDEO "videoSource -- failed to parse command line options\n");
-		return NULL;
-	}
-
-	return Create(opt);
-}
-
-// Create
-videoSource* videoSource::Create( const char* resource, const int argc, char** argv, int positionArg, const videoOptions& options )
-{
-	return Create(resource, commandLine(argc, argv), positionArg, options);
-}
-
-// Create
-videoSource* videoSource::Create( const commandLine& cmdLine, int positionArg )
-{
-	return Create(NULL, cmdLine, positionArg);
-}
-
-// Create
-videoSource* videoSource::Create( const int argc, char** argv, int positionArg )
-{
-	return Create(commandLine(argc, argv), positionArg);
-}
-
-// Create
-videoSource* videoSource::Create( const char* resource, const videoOptions& options )
-{
-	return Create(resource, 0, NULL, -1, options);
-}
-
 // Open
 bool videoSource::Open()
 {
@@ -131,15 +62,15 @@ void videoSource::Close()
 }
 
 // TypeToStr
-const char* videoSource::TypeToStr( uint32_t type )
+const char* videoSource::TypeToStr( uint32_t type ) const
 {
 	if( type == gstCamera::Type )
 		return "gstCamera";
-	else if( type == gstDecoder::Type )
-		return "gstDecoder";
-	else if( type == imageLoader::Type )
-		return "imageLoader";
-
+	// else if( type == gstDecoder::Type )
+	// 	return "gstDecoder";
+	// else if( type == imageLoader::Type )
+	// 	return "imageLoader";
+	assert(false);
 	return "(unknown)";
 }
